// src/api/models/field.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Burbot Business Driver
 * OpenAPI spec version: 0.1.0
 */
import type { SourceLocation } from './sourceLocation';
import type { Type } from './type';

export interface Field {
  name: string;
  type: Type;
  location: SourceLocation;
  nullable?: boolean;
  unique?: boolean;
  index?: boolean;
  primary_key?: boolean;
  default?: unknown | null;
  min_length?: number | null;
  max_length?: number | null;
  pattern?: string | null;
  gt?: number | null;
  ge?: number | null;
  lt?: number | null;
  le?: number | null;
  precision?: number | null;
  scale?: number | null;
  reference?: string | null;
}